requirements.txt
Flask==3.0.0
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
Flask-Talisman==1.1.0
Flask-Limiter==3.5.0
pymongo==4.6.1
motor==3.3.2
argon2-cffi==23.1.0
pyotp==2.9.0
qrcode==7.4.2
cryptography==41.0.7
# scikit-learn==1.3.2
# pandas==2.1.4
# numpy==1.26.2
weasyprint==60.2
pytest==7.4.3
pytest-cov==4.1.0
python-dotenv==1.0.0
email_validator==2.1.0.post1
bleach==6.1.0
